+++ components/employer-home-page.tsx
@@ -2,22 +2,21 @@
 
 import { Employee, Employer } from '@/types/class.service'
 import React, { useEffect, useState } from 'react'
-import { getDateOfApply } from '@/app/main/shifts-application/shifts_apply_table'
 import { Button } from '@nextui-org/react'
 import { useRouter } from 'next/navigation'
 import { getEmployeesByFilter } from '@/services/server-services/employer.service'
+import { getDateOfApply } from '@/lib/server.utils'
 
 export default function EmployeeHomePage({ employerUser }: { employerUser: Employer }) {
-  // Todo: Modal opens if the employer need to do something urgent 
-  const [forDate, setForDate] = useState<string>(undefined)
+
   const [usersNotApplied, setUsersNotApplied] = useState<Employee[]>([])
   const router = useRouter()
-
+  const forDate = getDateOfApply(employerUser.applicationTime.day, employerUser.applicationTime.time)
+  
   useEffect(() => {
-    setForDate(getDateOfApply(employerUser.applicationTime.day, employerUser.applicationTime.time))
 
     const getEmployees = async () => {
-      const data = await getEmployeesByFilter({field: 'isApplied', value: false}, employerUser.id)
+      const data = await getEmployeesByFilter({isApplied: false}, employerUser.id)
       setUsersNotApplied(data)
     }
     getEmployees()
